ui: separate wallet accounts modal
